requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
hypothesis==6.92.0

# Async & Web
fastapi==0.108.0
uvicorn==0.25.0
httpx==0.23.0
aiohttp==3.9.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1

# AI/ML
google-generativeai==0.3.2
pinecone-client==3.0.0

# Blockchain
solana==0.30.2
solders==0.18.1
anchorpy==0.18.0

# Data & Analytics
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
python-json-logger==2.0.7
pytz==2023.3
tenacity==8.2.3

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
pre-commit==3.6.0
